+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Unreleased template stuff
 - Added `torii.lib.stream.simple.ConstantStreamGenerator`, which is a port of the `torii_usb.stream.generator.ConstantStreamGenerator`.
 - Added `torii.lib.stream.simple.StreamSerializer`, which is a port of the `torii_usb.stream.generator.StreamSerializer`.
 - New WASM based simulation backend [torii-sim-wasm](https://github.com/shrine-maiden-heavy-industries/torii-sim-wasm) leveraging [wasmtime](https://wasmtime.dev/) to speed up larger and longer running simulation tests in a cross-platform manner.
+- New `torii.lib.lfsr` module for Galois and Fibonacci Linear-feedback shift registers (LFSRs).
 
 ### Changed